# fraud-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fraud scoring API",
    "description": "캡스톤 연구와 분리한 추론 서비스",
    "version": "0.2.0"
  },
  "paths": {
...
    "/admin/api/rules/{rule_id}/toggle": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Toggle Rule",
        "operationId": "admin_toggle_rule_admin_api_rules__rule_id__toggle_post",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/api/threshold": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Set Threshold",
        "description": "body 예시: {\"BLOCK_THRESHOLD\": 0.9, \"REVIEW_THRESHOLD\": 0.4}\nSYSTEM_CONFIG에 있는 키만 수정 허용.",
        "operationId": "admin_set_threshold_admin_api_threshold_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/profile/ingest": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Ingest",
        "operationId": "ingest_v1_profile_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profile/{user_id}": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_v1_profile__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Profile V1 Profile  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "Delete Profile",
        "operationId": "delete_profile_v1_profile__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "tx_id": {
            "type": "string",
            "title": "Tx Id",
            "default": ""
          },
          "amount": {
            "type": "number",
            "minimum": 0.0,
            "title": "Amount"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "merchant_id": {
            "type": "string",
            "title": "Merchant Id",
            "default": ""
          },
          "device_id": {
            "type": "string",
            "title": "Device Id",
            "default": ""
          },
          "ip": {
            "type": "string",
            "title": "Ip",
            "default": ""
          },
          "is_foreign_ip": {
            "type": "boolean",
            "title": "Is Foreign Ip",
            "default": false
          },
          "hour": {
            "type": "integer",
            "maximum": 23.0,
            "minimum": -1.0,
            "title": "Hour",
            "default": -1
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "amount"
        ],
        "title": "IngestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}